[ODA-229] Fix bug rolling back an Update Operation with deployment elements of type Install

+++ oda-operations/update/src/main/java/es/amplia/oda/operation/update/operations/InstallDeploymentElementOperation.java
@@ -28,6 +28,8 @@ public class InstallDeploymentElementOperation extends DeploymentElementOperatio
 
     @Override
     protected void rollbackSpecificOperation(FileManager fileManager, String backupFile) throws FileException {
-        fileManager.delete(installedFile);
+        if (installedFile != null) {
+            fileManager.delete(installedFile);
+        }
     }
 }
